Move windowing and segmenting functions to separate module
